+++ src/services/music.service.ts
@@ -1,70 +1,70 @@
 import mongoose from "mongoose";
-import musicModel from "../models/music.model";
-import albumModel from "../models/album.model";
-import IMusic from "../interfaces/music.interface";
-import * as dbs from "./database.service";
-import IAlbum from "../interfaces/album.interface";
+import MusicModel from "../models/music.model";
+import AlbumModel from "../models/album.model";
+import { IMusic, IPopulatedMusic } from "../interfaces/music.interface";
+import { NotFoundError } from "./errorHandling.service";
+import { IAlbum } from "../interfaces/album.interface";
+import IUser from "../interfaces/user.interface";
+
+// Populate music
+async function _populate(musicDoc: mongoose.Document<unknown, any, IMusic> & IMusic): Promise<IPopulatedMusic> {
+  const populatedDoc: IPopulatedMusic = await (await musicDoc.populate<{ authors: IUser[] }>("authors")).toObject();
+  populatedDoc.authors.forEach((author, index) => {
+    delete populatedDoc.authors[index].email;
+  });
+
+  return populatedDoc;
+}
 
 // Operação CREATE
-async function _create(MusicData: IMusic): Promise<IMusic> {
-  const newMusic = new musicModel(MusicData);
-  await dbs.validate(newMusic);
+async function _create(MusicData: IMusic): Promise<mongoose.Document<unknown, any, IMusic> & IMusic> {
+  const newMusic = new MusicModel(MusicData);
+  await newMusic.validate();
   await newMusic.save();
 
-  await dbs.redisSET(newMusic.id, JSON.stringify(newMusic));
-
-  return newMusic.toObject();
+  return newMusic;
 }
 
 // Operação READ
-async function _read(id: string): Promise<IMusic> {
-  let musicDoc: string | (mongoose.Document & IMusic) | null;
+async function _read(id: string): Promise<mongoose.Document<unknown, any, IMusic> & IMusic> {
+  const musicDoc = await MusicModel.findById(id);
+  if (!musicDoc) throw new NotFoundError("Music not found");
 
-  // Busca no Redis
-  musicDoc = await dbs.redisGET(id);
-  if (musicDoc) return JSON.parse(musicDoc);
-
-  musicDoc = await dbs.getDocumentById("MusicModel", id);
-  return musicDoc.toObject();
+  return musicDoc;
 }
 
 // Operação UPDATE
-async function _update(id: string, newData: IMusic): Promise<IMusic> {
-  let musicDoc: mongoose.Document & IMusic = await dbs.getDocumentById(
-    "MusicModel",
-    id
-  );
+async function _update(id: string, newData: IMusic): Promise<mongoose.Document<unknown, any, IMusic> & IMusic> {
+  const musicDoc = await MusicModel.findById(id);
+  if (!musicDoc) throw new NotFoundError("Music not found");
 
   if (newData.name) musicDoc.name = newData.name;
   if (newData.cover) musicDoc.cover = newData.cover;
   if (newData.authors) musicDoc.authors = newData.authors;
-  if (newData.genre) musicDoc.genre = newData.genre;
 
   // Verificação e atualização no database
-  await dbs.validate(musicDoc);
+  await musicDoc.validate();
   await musicDoc.save();
 
-  // Atualização no Redis
-  await dbs.redisSET(id, JSON.stringify(musicDoc));
-
-  return musicDoc.toObject();
+  return musicDoc;
 }
 
 // Operação DELETE
 async function _delete(id: string): Promise<void> {
-  let musicDoc = await dbs.getDocumentById("MusicModel", id);
-  await musicDoc.delete();
-  await dbs.redisDEL(id);
+  const musicDoc = await MusicModel.findById(id);
+  if (!musicDoc) throw new NotFoundError("Music not found");
 
+  await musicDoc.delete();
   return;
 }
 
-async function _getAlbum(id: string): Promise<IAlbum> {
-  const albumDoc = await albumModel.find({ musics: id }).lean();
-  return albumDoc as IAlbum;
+async function _getAlbum(id: string): Promise<mongoose.Document<unknown, any, IAlbum> & IAlbum | null> {
+  const albumDoc = await AlbumModel.findOne({ musics: id });
+  return albumDoc;
 }
 
 export = {
+  populate: _populate,
   create: _create,
   read: _read,
   update: _update,
